Type of id fields changed to long

+++ src/main/java/dao/CourseDAO.java
@@ -15,7 +15,7 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
     public void add(Course entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "insert into courses.courses values(?, ?, ?, ?, ?, ?)")) {
-            preparedStatement.setInt(1, entity.getId());
+            preparedStatement.setLong(1, entity.getId());
             preparedStatement.setString(2, entity.getName());
             preparedStatement.setDate(3, entity.getCreatedAt());
             preparedStatement.setTimestamp(4, entity.getStartDateTime());
@@ -28,16 +28,16 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
     }
 
     @Override
-    public Course getById(int id) {
+    public Course getById(long id) {
         Course course = null;
         try (Connection connection = HikariCPDataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(
                      "select * from courses.courses where id=?")) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 if (resultSet.next()) {
                     course = Course.builder()
-                            .id(resultSet.getInt(1))
+                            .id(resultSet.getLong(1))
                             .name(resultSet.getString(2))
                             .createdAt(resultSet.getDate(3))
                             .startDateTime(resultSet.getTimestamp(4))
@@ -61,7 +61,7 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
              ResultSet resultSet = preparedStatement.executeQuery()) {
             while (resultSet.next()) {
                 courses.add(Course.builder()
-                        .id(resultSet.getInt(1))
+                        .id(resultSet.getLong(1))
                         .name(resultSet.getString(2))
                         .createdAt(resultSet.getDate(3))
                         .startDateTime(resultSet.getTimestamp(4))
@@ -81,7 +81,7 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
                      "update courses.courses " +
                              "set name=?, created_at=? ,start_datetime=?, end_datetime=?, status=? " +
                              "where id=?")) {
-            preparedStatement.setInt(6, entity.getId());
+            preparedStatement.setLong(6, entity.getId());
             preparedStatement.setString(1, entity.getName());
             preparedStatement.setDate(2, entity.getCreatedAt());
             preparedStatement.setTimestamp(3, entity.getStartDateTime());
@@ -97,7 +97,7 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
     public void remove(Course entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "delete from courses.courses where id=?")) {
-            preparedStatement.setInt(1, entity.getId());
+            preparedStatement.setLong(1, entity.getId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -113,7 +113,7 @@ public class CourseDAO implements DAOConnectionPassing<Course> {
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
                     courses.add(Course.builder()
-                            .id(resultSet.getInt(1))
+                            .id(resultSet.getLong(1))
                             .name(resultSet.getString(2))
                             .createdAt(resultSet.getDate(3))
                             .startDateTime(resultSet.getTimestamp(4))
