+++ src/main/java/dao/CreatedCourseDAO.java
@@ -15,8 +15,8 @@ public class CreatedCourseDAO implements DAOConnectionPassing<CreatedCourse> {
     public void add(CreatedCourse entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "insert into courses.created_courses values(?, ?)")) {
-            preparedStatement.setInt(1, entity.getUserId());
-            preparedStatement.setInt(2, entity.getCourseId());
+            preparedStatement.setLong(1, entity.getUserId());
+            preparedStatement.setLong(2, entity.getCourseId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -24,17 +24,17 @@ public class CreatedCourseDAO implements DAOConnectionPassing<CreatedCourse> {
     }
 
     @Override
-    public CreatedCourse getById(int id) {
+    public CreatedCourse getById(long id) {
         CreatedCourse createdCourse = null;
         try (Connection connection = HikariCPDataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(
                      "select * from courses.created_courses where user_id=?")) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 if(resultSet.next()) {
                     createdCourse = CreatedCourse.builder()
-                            .userId(resultSet.getInt(1))
-                            .courseId(resultSet.getInt(2))
+                            .userId(resultSet.getLong(1))
+                            .courseId(resultSet.getLong(2))
                             .build();
                 }
             }
@@ -53,8 +53,8 @@ public class CreatedCourseDAO implements DAOConnectionPassing<CreatedCourse> {
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
                     createdCourses.add(CreatedCourse.builder()
-                            .userId(resultSet.getInt(1))
-                            .courseId(resultSet.getInt(2))
+                            .userId(resultSet.getLong(1))
+                            .courseId(resultSet.getLong(2))
                             .build());
                 }
             }
@@ -68,10 +68,10 @@ public class CreatedCourseDAO implements DAOConnectionPassing<CreatedCourse> {
     public void edit(CreatedCourse entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "update courses.created_courses set user_id=?, course_id=? where user_id=? and course_id=?")) {
-            preparedStatement.setInt(1, entity.getUserId());
-            preparedStatement.setInt(2, entity.getCourseId());
-            preparedStatement.setInt(3, entity.getUserId());
-            preparedStatement.setInt(4, entity.getCourseId());
+            preparedStatement.setLong(1, entity.getUserId());
+            preparedStatement.setLong(2, entity.getCourseId());
+            preparedStatement.setLong(3, entity.getUserId());
+            preparedStatement.setLong(4, entity.getCourseId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -82,8 +82,8 @@ public class CreatedCourseDAO implements DAOConnectionPassing<CreatedCourse> {
     public void remove(CreatedCourse entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "delete from courses.created_courses where user_id=? and course_id=?")) {
-            preparedStatement.setInt(1, entity.getUserId());
-            preparedStatement.setInt(2, entity.getCourseId());
+            preparedStatement.setLong(1, entity.getUserId());
+            preparedStatement.setLong(2, entity.getCourseId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
