+++ src/main/java/dao/DAOConnectionPassing.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface DAOConnectionPassing<E> {
     void add(E entity, Connection connection);
-    E getById(int id);
+    E getById(long id);
     List<E> getAll();
     void edit(E entity, Connection connection);
     void remove(E entity, Connection connection);
