+++ src/main/java/dao/StudyDAO.java
@@ -15,9 +15,9 @@ public class StudyDAO implements DAOConnectionPassing<Study> {
     public void add(Study entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "insert into courses.study values(?, ?, ?)")) {
-            preparedStatement.setInt(1, entity.getId());
-            preparedStatement.setInt(2, entity.getCourseId());
-            preparedStatement.setInt(3, entity.getUserId());
+            preparedStatement.setLong(1, entity.getId());
+            preparedStatement.setLong(2, entity.getCourseId());
+            preparedStatement.setLong(3, entity.getUserId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -25,18 +25,18 @@ public class StudyDAO implements DAOConnectionPassing<Study> {
     }
 
     @Override
-    public Study getById(int id) {
+    public Study getById(long id) {
         Study study = null;
         try (Connection connection = HikariCPDataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(
                      "select * from courses.study where id=?")) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 if(resultSet.next()) {
                     study = Study.builder()
-                            .id(resultSet.getInt(1))
-                            .courseId(resultSet.getInt(2))
-                            .userId(resultSet.getInt(3))
+                            .id(resultSet.getLong(1))
+                            .courseId(resultSet.getLong(2))
+                            .userId(resultSet.getLong(3))
                             .build();
                 }
             }
@@ -55,9 +55,9 @@ public class StudyDAO implements DAOConnectionPassing<Study> {
             try (ResultSet resultSet = preparedStatement.executeQuery()) {
                 while (resultSet.next()) {
                     studies.add(Study.builder()
-                            .id(resultSet.getInt(1))
-                            .courseId(resultSet.getInt(2))
-                            .userId(resultSet.getInt(3))
+                            .id(resultSet.getLong(1))
+                            .courseId(resultSet.getLong(2))
+                            .userId(resultSet.getLong(3))
                             .build());
                 }
             }
@@ -71,9 +71,9 @@ public class StudyDAO implements DAOConnectionPassing<Study> {
     public void edit(Study entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "update courses.study set user_id=?, course_id=? where id=?")) {
-            preparedStatement.setInt(1, entity.getUserId());
-            preparedStatement.setInt(2, entity.getCourseId());
-            preparedStatement.setInt(3, entity.getId());
+            preparedStatement.setLong(1, entity.getUserId());
+            preparedStatement.setLong(2, entity.getCourseId());
+            preparedStatement.setLong(3, entity.getId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -84,7 +84,7 @@ public class StudyDAO implements DAOConnectionPassing<Study> {
     public void remove(Study entity, Connection connection) {
         try (PreparedStatement preparedStatement = connection.prepareStatement(
                      "delete from courses.study where id=?")) {
-            preparedStatement.setInt(1, entity.getId());
+            preparedStatement.setLong(1, entity.getId());
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
